refactor: move job-queue to node-app
