+++ imports/node-app/plugins/job-queue/jobs.js
@@ -1,5 +1,5 @@
 import later from "later";
-import createJobCollection from "./job-queue/createJobCollection";
+import createJobCollection from "./job-queue/createJobCollection.js";
 
 const { Job, JobCollection } = createJobCollection({ later });
