+++ imports/plugins/included/job-queue/register.js
@@ -6,6 +6,6 @@
  * temporary bridge.
  */
 import Reaction from "/imports/plugins/core/core/server/Reaction";
-import register from "./server/no-meteor/register";
+import register from "/imports/node-app/plugins/job-queue/index.js";
 
 Reaction.whenAppInstanceReady(register);
